+++ CHANGES.rst
@@ -172,6 +172,7 @@ Unreleased
     addition to its type. :issue:`457`
 -   ``confirmation_prompt`` can be set to a custom string. :issue:`723`
 -   Allow styled output in Jupyter on Windows. :issue:`1271`
+-   ``style()`` supports the ``strikethrough`` style. :issue:`805`
 
 
 Version 7.1.2